services/backend/app/requirements.txt
aio-pika==9.4.2
aiofiles==23.2.1
alembic==1.12.1
alembic-postgresql-enum==1.3.0
fastapi==0.104.1
polars==0.20.7
Pillow==10.2.0
psycopg2-binary==2.9.9
pydantic[dotenv]==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
pyjwt[crypto]==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.5
requests==2.27.1
sqlalchemy==2.0.23
uvicorn==0.16.0
boto3==1.34.31
httpx==0.27.0
SQLAlchemy==2.0.23
sse-starlette==2.1.2
passlib==1.7.4
python-keycloak==4.3.0
bcrypt==4.2.0